+++ CHANGES.md
@@ -8,6 +8,7 @@
   `internalSubset` properties
 - Fix: Throw proper `Error` on undefined child content
 - Fix: Allow previously undocumented fragment creation to accept children
+- Fix: Throw for all non-object plugins
 - Linting (ESLint): As per latest ash-nazg
 - Travis: Adding builds 10, 12
 - Testing: Coverage